Add separate fastBlur and warpsharp shaders
A big and fast blur. Only 4 passes, each with 5 texture reads.

Warp sharp sharpens blurry edges by detecting edges then sampling a point away from the centre line of the edge.
	
Warp sharp is good for restoring clear edges (but not texture details) on images that have been blurred or upscaled.
	
It is not necessarily good for sharpening game output, as games have a mixture of very sharp edges of objects, and varying levels of blurrines in different textures.
	
This sharpening algorithm is based roughly on the description here: https://www.virtualdub.org/blog2/entry_079.html - except I have addressed the contrast problem describe there by taking the square root of the bump map value. Increasing the scale is fast but not as clever as it could be - ideally you'd use a bigger kernel at higher scales.